pyproject.toml
[project]
name = "project-dbt-bigquery"
version = "0.1.0"
description = ""
authors = [
    {name = "aletheavilla"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dbt-core (>=1.9.2,<2.0.0)",
    "dbt-bigquery (>=1.9.1,<2.0.0)",
    "google-cloud-bigquery (>=3.30.0,<4.0.0)",
    "pandas (>=2.2.3,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
